+++ test/serialize.ts
@@ -99,3 +99,53 @@ test('attribute selector with non-empty namespace', serializeMacro,
   '[foo|attr*=value]',
   '[foo|attr*="value"]',
 );
+
+test('simple pseudo-class', serializeMacro,
+  ':hover',
+  ':hover',
+);
+
+test(':is() with multiple selectors', serializeMacro,
+  ':is( .foo , .bar , .baz )',
+  ':is(.foo,.bar,.baz)',
+);
+
+test(':where() with multiple selectors', serializeMacro,
+  ':where( #foo , #bar )',
+  ':where(#foo,#bar)',
+);
+
+test(':not() with multiple selectors', serializeMacro,
+  ':not( [disabled] , .hidden )',
+  ':not([disabled],.hidden)',
+);
+
+test('nested :not() inside :where() inside :is()', serializeMacro,
+  ':is(:where(:not(#foo)))',
+  ':is(:where(:not(#foo)))',
+);
+
+test('pseudo-classes in compound selector', serializeMacro,
+  'div.foo:hover:focus',
+  'div.foo:hover:focus',
+);
+
+test('functional pseudo-classes in compound selector', serializeMacro,
+  'div:is(.a, .b):not(.c)',
+  'div:is(.a,.b):not(.c)',
+);
+
+test('functional pseudo-class with combinators inside', serializeMacro,
+  ':is(div > p, span + a)',
+  ':is(div>p,span+a)',
+);
+
+test('functional pseudo-class with combinator outside', serializeMacro,
+  'div:is(.foo) > span',
+  'div:is(.foo)>span',
+);
+
+test('text case of various selectors', serializeMacro,
+  'P.CLASS#ID[ATTR="VALUE"]:HOVER:IS(.FOO):WHERE(.BAR):NOT(.BAZ)',
+  'P.CLASS#ID[ATTR="VALUE"]:HOVER:IS(.FOO):WHERE(.BAR):NOT(.BAZ)',
+);